Move key operations to separate module
